requirements.txt
requests
feedparser
feedsearch
pygooglenews
newspaper3k
fastapi
uvicorn
jinja2
python-dateutil
